+++ src/OpenTelemetry.AutoInstrumentation/Loading/Initializers/AspNetMetricsInitializer.cs
@@ -3,14 +3,18 @@
 
 #if NETFRAMEWORK
 
+using OpenTelemetry.AutoInstrumentation.Plugins;
+
 namespace OpenTelemetry.AutoInstrumentation.Loading.Initializers;
 
 internal class AspNetMetricsInitializer
 {
+    private readonly PluginManager _pluginManager;
     private int _initialized;
 
-    public AspNetMetricsInitializer(LazyInstrumentationLoader lazyInstrumentationLoader)
+    public AspNetMetricsInitializer(LazyInstrumentationLoader lazyInstrumentationLoader, PluginManager pluginManager)
     {
+        _pluginManager = pluginManager;
         lazyInstrumentationLoader.Add(new AspNetMvcInitializer(InitializeOnFirstCall));
         lazyInstrumentationLoader.Add(new AspNetWebApiInitializer(InitializeOnFirstCall));
     }
@@ -24,7 +28,11 @@ internal class AspNetMetricsInitializer
         }
 
         var instrumentationType = Type.GetType("OpenTelemetry.Instrumentation.AspNet.AspNetMetrics, OpenTelemetry.Instrumentation.AspNet");
-        var instrumentation = Activator.CreateInstance(instrumentationType);
+
+        var options = new OpenTelemetry.Instrumentation.AspNet.AspNetMetricsInstrumentationOptions();
+        _pluginManager.ConfigureTracesOptions(options);
+
+        var instrumentation = Activator.CreateInstance(instrumentationType, args: options);
 
         lifespanManager.Track(instrumentation);
     }